Attribute Macro rithmic_impl::autofill

#[autofill]

Create a wrapper macro with the same name as a given function, which automatically fills some of the later arguments with variables of the same name in the calling scope. The result is similar to a closure’s external variable capture but without the disadvantages of closures (e.g. lack of recursion, borrow conflicts)

Note: #![feature(rustc_attrs)] is required

Examples

#![feature(rustc_attrs)]

use rithmic::autofill;

let mut u = [3, 5, 7, 11, 13];
let mut v = vec![];

#[autofill(2..)]
fn duplicate(i: usize, count: usize, u: &[i32], v: &mut Vec<i32>)
{
    if count != 0 {
        v.push(u[i]);
        duplicate!(i, count - 1);  // note the !
    }
}

duplicate!(2, 4);

assert_eq!(v, vec![7, 7, 7, 7]);