Attribute Macro rithmic_impl::memoize

#[memoize]

Memoize a function based on some or all of its arguments

Two types of storage are supported:

Static: #[memoize[5, 7]] uses an NdVec of size 5 x 7 to memoize based on the first 2 function arguments, which must be castable to usize

Dynamic: #[memoize(..3)] uses an AHashMap to memoize based on the first 3 function arguments, which must be hashable

A wrapper macro with the same name as the function is created, and this macro must be used instead of calling the function directly. Arguments other than the memoized parameters are auto-filled with names from the calling scope; see autofill

Note: #![feature(rustc_attrs)] is required

Examples

Static

#![feature(rustc_attrs)]

use rithmic::memoize;

#[memoize[10]]
fn fibonacci(n: usize) -> u64 {
    match n {
        0 => 0,
        1 => 1,
        _ => fibonacci!(n-1) + fibonacci!(n-2)
    }
}

assert_eq!(fibonacci!(9), 34);

Dynamic

#![feature(rustc_attrs)]

use rithmic::memoize;

#[memoize(..1)]
fn fibonacci(n: usize) -> u64 {
    match n {
        0 => 0,
        1 => 1,
        _ => fibonacci!(n-1) + fibonacci!(n-2)
    }
}

assert_eq!(fibonacci!(20), 6765);