Crate rithmic_impl

Macros

struct_input

Given a ::proconio input specification and a Name, define:

Attribute Macros

autofill

Create a wrapper macro with the same name as a given function, which automatically fills some of the later arguments with variables of the same name in the calling scope. The result is similar to a closure’s external variable capture but without the disadvantages of closures (e.g. lack of recursion, borrow conflicts)

memoize

Memoize a function based on some or all of its arguments

opaque

See transparent

semitransparent

See transparent

transparent

Change the hygiene of a declarative macro

Derive Macros

CmpByKey

Derive PartialEq, Eq, PartialOrd and Ord for a struct, but only based on a subset of its fields, marked by #[key]